feat: Add user_object_from_credentials() method to BasicAuth class
The user_object_from_credentials() method is added to the BasicAuth class to retrieve a User object based on user credentials. This method takes a user email and password as input and returns the corresponding User object if found. If the user is not found or the credentials are incorrect, None is returned.

This commit builds upon previous commits that added functionality to decode the base64 authorization header and extract user credentials. It also includes styling fixes and documentation updates.

Note: The commit message has been modified to adhere to the established commit message conventions.