feat: Add session_cookie() method to Auth class
The session_cookie() method is added to the Auth class to retrieve the session cookie from the request. This method checks if the request is None and returns None if it is. Otherwise, it gets the session name from the environment variable 'SESSION_NAME' and returns the corresponding cookie value from the request.